frontend/requirements.txt
gradio==4.*
sseclient-py==1.7      # newer version fixes bytes/str issue
requests>=2.31
